+++ contracts/IPoolBuilder.sol
@@ -8,8 +8,8 @@ interface IPoolBuilder {
         address _derivativeVault,
         address _feeCalculator,
         address _repricer,
-        uint _baseFee,
-        uint _maxFee,
-        uint _feeAmp
-    ) external returns(address);
+        uint256 _baseFee,
+        uint256 _maxFee,
+        uint256 _feeAmp
+    ) external returns (address);
 }
