+++ contracts/Num.sol
@@ -13,45 +13,29 @@
 
 pragma solidity 0.7.6;
 
-import "./Const.sol";
+import './Const.sol';
 
 contract Num is Const {
-
-    function toi(uint a)
-        internal pure
-        returns (uint)
-    {
+    function toi(uint256 a) internal pure returns (uint256) {
         return a / BONE;
     }
 
-    function floor(uint a)
-        internal pure
-        returns (uint)
-    {
+    function floor(uint256 a) internal pure returns (uint256) {
         return toi(a) * BONE;
     }
 
-    function add(uint a, uint b)
-        internal pure
-        returns (uint c)
-    {
+    function add(uint256 a, uint256 b) internal pure returns (uint256 c) {
         c = a + b;
-        require(c >= a, "ADD_OVERFLOW");
+        require(c >= a, 'ADD_OVERFLOW');
     }
 
-    function sub(uint a, uint b)
-        internal pure
-        returns (uint c)
-    {
+    function sub(uint256 a, uint256 b) internal pure returns (uint256 c) {
         bool flag;
         (c, flag) = subSign(a, b);
-        require(!flag, "SUB_UNDERFLOW");
+        require(!flag, 'SUB_UNDERFLOW');
     }
 
-    function subSign(uint a, uint b)
-        internal pure
-        returns (uint, bool)
-    {
+    function subSign(uint256 a, uint256 b) internal pure returns (uint256, bool) {
         if (a >= b) {
             return (a - b, false);
         } else {
@@ -59,34 +43,25 @@ contract Num is Const {
         }
     }
 
-    function mul(uint a, uint b)
-        internal pure
-        returns (uint c)
-    {
-        uint c0 = a * b;
-        require(a == 0 || c0 / a == b, "MUL_OVERFLOW");
-        uint c1 = c0 + (BONE / 2);
-        require(c1 >= c0, "MUL_OVERFLOW");
+    function mul(uint256 a, uint256 b) internal pure returns (uint256 c) {
+        uint256 c0 = a * b;
+        require(a == 0 || c0 / a == b, 'MUL_OVERFLOW');
+        uint256 c1 = c0 + (BONE / 2);
+        require(c1 >= c0, 'MUL_OVERFLOW');
         c = c1 / BONE;
     }
 
-    function div(uint a, uint b)
-        internal pure
-        returns (uint c)
-    {
-        require(b != 0, "DIV_ZERO");
-        uint c0 = a * BONE;
-        require(a == 0 || c0 / a == BONE, "DIV_INTERNAL"); // mul overflow
-        uint c1 = c0 + (b / 2);
-        require(c1 >= c0, "DIV_INTERNAL"); //  add require
+    function div(uint256 a, uint256 b) internal pure returns (uint256 c) {
+        require(b != 0, 'DIV_ZERO');
+        uint256 c0 = a * BONE;
+        require(a == 0 || c0 / a == BONE, 'DIV_INTERNAL'); // mul overflow
+        uint256 c1 = c0 + (b / 2);
+        require(c1 >= c0, 'DIV_INTERNAL'); //  add require
         c = c1 / b;
     }
 
     // DSMath.wpow
-    function powi(uint a, uint n)
-        internal pure
-        returns (uint z)
-    {
+    function powi(uint256 a, uint256 n) internal pure returns (uint256 z) {
         z = n % 2 != 0 ? a : BONE;
 
         for (n /= 2; n != 0; n /= 2) {
@@ -101,45 +76,42 @@ contract Num is Const {
     // Compute b^(e.w) by splitting it into (b^e)*(b^0.w).
     // Use `powi` for `b^e` and `powK` for k iterations
     // of approximation of b^0.w
-    function pow(uint base, uint exp)
-        internal pure
-        returns (uint)
-    {
-        require(base >= MIN_POW_BASE, "POW_BASE_TOO_LOW");
-        require(base <= MAX_POW_BASE, "POW_BASE_TOO_HIGH");
+    function pow(uint256 base, uint256 exp) internal pure returns (uint256) {
+        require(base >= MIN_POW_BASE, 'POW_BASE_TOO_LOW');
+        require(base <= MAX_POW_BASE, 'POW_BASE_TOO_HIGH');
 
-        uint whole  = floor(exp);
-        uint remain = sub(exp, whole);
+        uint256 whole = floor(exp);
+        uint256 remain = sub(exp, whole);
 
-        uint wholePow = powi(base, toi(whole));
+        uint256 wholePow = powi(base, toi(whole));
 
         if (remain == 0) {
             return wholePow;
         }
 
-        uint partialResult = powApprox(base, remain, POW_PRECISION);
+        uint256 partialResult = powApprox(base, remain, POW_PRECISION);
         return mul(wholePow, partialResult);
     }
 
-    function powApprox(uint base, uint exp, uint precision)
-        internal pure
-        returns (uint sum)
-    {
+    function powApprox(
+        uint256 base,
+        uint256 exp,
+        uint256 precision
+    ) internal pure returns (uint256 sum) {
         // term 0:
-        uint a     = exp;
-        (uint x, bool xneg)  = subSign(base, BONE);
-        uint term = BONE;
-        sum   = term;
+        uint256 a = exp;
+        (uint256 x, bool xneg) = subSign(base, BONE);
+        uint256 term = BONE;
+        sum = term;
         bool negative = false;
-
 
         // term(k) = numer / denom
         //         = (product(a - i - 1, i=1-->k) * x^k) / (k!)
         // each iteration, multiply previous term by (a-(k-1)) * x / k
         // continue until term is less than precision
-        for (uint i = 1; term >= precision; i++) {
-            uint bigK = i * BONE;
-            (uint c, bool cneg) = subSign(a, sub(bigK, BONE));
+        for (uint256 i = 1; term >= precision; i++) {
+            uint256 bigK = i * BONE;
+            (uint256 c, bool cneg) = subSign(a, sub(bigK, BONE));
             term = mul(term, mul(c, x));
             term = div(term, bigK);
             if (term == 0) break;
@@ -154,11 +126,8 @@ contract Num is Const {
         }
     }
 
-    function min(uint first, uint second)
-        internal pure
-        returns (uint)
-    {
-        if(first < second) {
+    function min(uint256 first, uint256 second) internal pure returns (uint256) {
+        if (first < second) {
             return first;
         }
         return second;
