+++ contracts/repricers/IRepricer.sol
@@ -2,23 +2,27 @@
 
 pragma solidity 0.7.6;
 
-import "../IVault.sol";
+import '../libs/complifi/IVault.sol';
 
 interface IRepricer {
-
-    function isRepricer() external pure returns(bool);
+    function isRepricer() external pure returns (bool);
 
     function symbol() external pure returns (string memory);
 
     function reprice(
-        uint _pMin,
-        int _volatility,
+        uint256 _pMin,
+        int256 _volatility,
         IVault _vault,
-        uint[2] memory _primary,
-        uint[2] memory _complement,
-        int _liveUnderlingValue
+        uint256[2] memory _primary,
+        uint256[2] memory _complement,
+        int256 _liveUnderlingValue
     )
-    external view returns(
-        uint newPrimaryLeverage, uint newComplementLeverage, int estPricePrimary, int estPriceComplement
-    );
+        external
+        view
+        returns (
+            uint256 newPrimaryLeverage,
+            uint256 newComplementLeverage,
+            int256 estPricePrimary,
+            int256 estPriceComplement
+        );
 }
