+++ contracts/test/TToken.sol
@@ -16,31 +16,31 @@ pragma solidity 0.7.6;
 // Test Token
 
 contract TToken {
-
     string private _name;
     string private _symbol;
-    uint8   private _decimals;
+    uint8 private _decimals;
 
     address private _owner;
 
-    uint internal _totalSupply;
+    uint256 internal _totalSupply;
 
-    mapping(address => uint)                   private _balance;
-    mapping(address => mapping(address=>uint)) private _allowance;
+    mapping(address => uint256) private _balance;
+    mapping(address => mapping(address => uint256)) private _allowance;
 
     modifier _onlyOwner_() {
-        require(msg.sender == _owner, "NOT_OWNER");
+        require(msg.sender == _owner, 'NOT_OWNER');
         _;
     }
 
-    event Approval(address indexed src, address indexed dst, uint amt);
-    event Transfer(address indexed src, address indexed dst, uint amt);
+    event Approval(address indexed src, address indexed dst, uint256 amt);
+    event Transfer(address indexed src, address indexed dst, uint256 amt);
 
     // Math
-    function add(uint a, uint b) internal pure returns (uint c) {
+    function add(uint256 a, uint256 b) internal pure returns (uint256 c) {
         require((c = a + b) >= a);
     }
-    function sub(uint a, uint b) internal pure returns (uint c) {
+
+    function sub(uint256 a, uint256 b) internal pure returns (uint256 c) {
         require((c = a - b) <= a);
     }
 
@@ -63,44 +63,48 @@ contract TToken {
         return _symbol;
     }
 
-    function decimals() public view returns(uint8) {
+    function decimals() public view returns (uint8) {
         return _decimals;
     }
 
-    function _move(address src, address dst, uint amt) internal {
-        require(_balance[src] >= amt, "INSUFFICIENT_BAL");
+    function _move(
+        address src,
+        address dst,
+        uint256 amt
+    ) internal {
+        require(_balance[src] >= amt, 'INSUFFICIENT_BAL');
         _balance[src] = sub(_balance[src], amt);
         _balance[dst] = add(_balance[dst], amt);
         emit Transfer(src, dst, amt);
     }
 
-    function _push(address to, uint amt) internal {
+    function _push(address to, uint256 amt) internal {
         _move(address(this), to, amt);
     }
 
-    function _pull(address from, uint amt) internal {
+    function _pull(address from, uint256 amt) internal {
         _move(from, address(this), amt);
     }
 
-    function _mint(address dst, uint amt) internal {
+    function _mint(address dst, uint256 amt) internal {
         _balance[dst] = add(_balance[dst], amt);
         _totalSupply = add(_totalSupply, amt);
         emit Transfer(address(0), dst, amt);
     }
 
-    function allowance(address src, address dst) external view returns (uint) {
+    function allowance(address src, address dst) external view returns (uint256) {
         return _allowance[src][dst];
     }
 
-    function balanceOf(address whom) external view returns (uint) {
+    function balanceOf(address whom) external view returns (uint256) {
         return _balance[whom];
     }
 
-    function totalSupply() public view returns (uint) {
+    function totalSupply() public view returns (uint256) {
         return _totalSupply;
     }
 
-    function approve(address dst, uint amt) external returns (bool) {
+    function approve(address dst, uint256 amt) external returns (bool) {
         _allowance[msg.sender][dst] = amt;
         emit Approval(msg.sender, dst, amt);
         return true;
@@ -111,21 +115,25 @@ contract TToken {
         return true;
     }
 
-    function burn(uint amt) public returns (bool) {
-        require(_balance[address(this)] >= amt, "INSUFFICIENT_BAL");
+    function burn(uint256 amt) public returns (bool) {
+        require(_balance[address(this)] >= amt, 'INSUFFICIENT_BAL');
         _balance[address(this)] = sub(_balance[address(this)], amt);
         _totalSupply = sub(_totalSupply, amt);
         emit Transfer(address(this), address(0), amt);
         return true;
     }
 
-    function transfer(address dst, uint amt) external returns (bool) {
+    function transfer(address dst, uint256 amt) external returns (bool) {
         _move(msg.sender, dst, amt);
         return true;
     }
 
-    function transferFrom(address src, address dst, uint amt) external returns (bool) {
-        require(msg.sender == src || amt <= _allowance[src][msg.sender], "TOKEN_BAD_CALLER");
+    function transferFrom(
+        address src,
+        address dst,
+        uint256 amt
+    ) external returns (bool) {
+        require(msg.sender == src || amt <= _allowance[src][msg.sender], 'TOKEN_BAD_CALLER');
         _move(src, dst, amt);
         if (msg.sender != src && _allowance[src][msg.sender] != uint256(-1)) {
             _allowance[src][msg.sender] = sub(_allowance[src][msg.sender], amt);
